requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
h5py>=3.8
PyYAML>=6.0
scikit-learn>=1.3
torch>=2.1
jupyter>=1.0
ipykernel>=6.0
